feat: 통계 api 개발중

+++ src/main/java/com/example/egobook_be/domain/ego_room/dto/DayCountDto.java
@@ -2,6 +2,6 @@ package com.example.egobook_be.domain.ego_room.dto;
 
 public record DayCountDto(
         String day,
-        long count
+        Integer count
 ) {
 }
