+++ src/main/java/com/example/egobook_be/domain/ego_room/dto/HourCountDto.java
@@ -2,6 +2,6 @@ package com.example.egobook_be.domain.ego_room.dto;
 
 public record HourCountDto(
         int hour,
-        long count
+        Integer count
 ) {
 }
